DXF: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31617 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/dxf/ogrdxfdriver.cpp
@@ -72,7 +72,7 @@ static int OGRDXFDriverIdentify( GDALOpenInfo* poOpenInfo )
         i ++;
     if( !STARTS_WITH_CI(pszIter + i, "SECTION") )
         return FALSE;
-    i += strlen("SECTION");
+    i += static_cast<int>(strlen("SECTION"));
     return pszIter[i] == '\n' || pszIter[i] == '\r';
 }
